// src/types/chat.type.ts
export interface InitialChannelType {
  id?: string;
  name?: string;
  type?: string;
  profileId?: string;
  serverId?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface ChannelMemberType {
  id?: string;
  role?: string;
  profileId?: string;
  serverId?: string;
  createdAt?: string;
  updatedAt?: string;
}